+++ serialization/src/commonTest/kotlin/nl/adaptivity/xml/serialization/regressions/TestSoapHelper.kt
@@ -27,7 +27,6 @@ import nl.adaptivity.xmlutil.core.impl.multiplatform.use
 import nl.adaptivity.xmlutil.dom2.*
 import nl.adaptivity.xmlutil.serialization.XML
 import nl.adaptivity.xmlutil.util.CompactFragment
-import nl.adaptivity.xmlutil.util.CompactFragmentSerializer
 import kotlin.test.*